+++ batch/src/Components/Content.js
@@ -3,8 +3,8 @@ import Site from './Site';
 import Equipment from './Equipment';
 import { Route, Switch } from 'react-router-dom';
 import LoginCard from './LoginCard';
-import Form from './Form';
-import CheckList from './FormCheckList';
+import FormInput from './FormInput';
+import FormContent from './FormContent';
 import Callback from './Callback';
 import SettingsCard from './SettingsCard';
 import { formData, formTitle } from '../Data/FormData';
@@ -36,11 +36,11 @@ export default class Content extends Component {
             <div>
               <Route exact path='/' component={Site} />
               <Route path='/equip/:id' component={Equipment} />
-              <Route path='/form/:id' render={(props) => {
-                return <Form title={formTitle} formStructure={formData} {...props} />
+              <Route path='/FormInput/:id' render={(props) => {
+                return <FormInput {...props} title={formTitle} formStructure={formData} period='1'/>
               }} />
-              <Route path='/checklist/:id' render={(props) => {
-                return <CheckList title={formTitle} formStructure={formData} {...props} />
+              <Route path='/FormContent/:id' render={(props) => {
+                return <FormContent {...props} title={formTitle} formStructure={formData} period='3' />
               }} />
               <Route path='/settings' component={SettingsCard} />
             </div>
